+++ taf/hosts.py
@@ -139,7 +139,10 @@ def set_hosts_of_repo(auth_repo, hosts):
 def load_hosts_json(auth_repo, commit=None):
     if commit is None:
         commit = auth_repo.top_commit_of_branch(auth_repo.default_branch)
-    return _get_json_file(auth_repo, HOSTS_JSON_PATH, commit)
+    try:
+        return _get_json_file(auth_repo, HOSTS_JSON_PATH, commit)
+    except MissingHostsError:
+        return {}
 
 
 def _get_json_file(auth_repo, path, commit):